Add Bin Script to Run AWS File Upload
Change the S3 folder directory to organize all projects files within
a single folder.

Add bin script that takes a filename as a command line argument, and
runs the lib script that uploads to Amazon.

Fix an issue with Config variables missing when testing and debugging
the code.

Code successfully debugged and tested with images, text, and video files
.

Packages added:
  -Dotenv

Pair Programmers: Aaron and Vitorio